Moved private method to own module.
